requirements.txt
Django~=4.1.6
django-environ~=0.9.0
psycopg2-binary
gunicorn~=20.1.0
djangorestframework~=3.14.0
django-extensions~=3.2.1
social_auth_app_django~=5.0.0
django-filter~=22.1
pydantic~=1.10.6
requests~=2.28.2
